+++ CHANGES.md
@@ -4,6 +4,9 @@ Expyriment Release Notes
  
 upcoming version
 -----------------
+- new method: all visual stimuli have methods for lowlevel Pygame operations
+  get_surface_copy(), set_surface(), get_pixel_array()
+ 
 
 Version 0.8.0
 -------------